requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorboard>=2.14.0
tqdm>=4.65.0
Pillow>=10.0.0
PyYAML>=6.0
omegaconf>=2.3.0
matplotlib>=3.7.0
